Convert cam_control to a python module.
To make the control daemon installable via python's setuptools
we need to make it a module. This means that doing setup.py install
will now also update the API.

The module is named camControl, so the API can be run by doing
'python3 -m camControl' or just invoking /usr/bin/camControl.
