+++ camControl/__init__.py
@@ -1,0 +1,1 @@
+from camControl.api import controlApi
