+++ camControl/api.py
@@ -1,9 +1,6 @@
 #!/usr/bin/python3
-
 import os, sys, pdb
-import signal
 import json
-import argparse
 from functools import lru_cache
 
 import inspect
@@ -23,9 +20,8 @@ import pychronos.regmaps as regmaps
 
 from pychronos.utils import getBoardRevision
 
+# The D-Bus interface we are exporting.
 interface = 'ca.krontech.chronos.control'
-dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
-bus = dbus.SystemBus()
 
 #-----------------------------------------------------------------
 # Some constants that ought to go into a board-specific dict.
@@ -635,89 +631,4 @@ class controlApi(dbus.service.Object):
             return {
                 "error": "Invalid Resolution"
             }
-
-class cameraTimer:
-    def __init__(self, camera):
-        self.camera = camera
-
-    def __call__(self, *args):
-        self.camera.tick()
-        return True
-
-# Run the control API
-if __name__ == "__main__":
-    # Enable logging.
-    logging.basicConfig(level=logging.DEBUG, format='%(asctime)s %(levelname)s [%(funcName)s] %(message)s')
-    
-    # Default sensors to use by top byte of board revision.
-    sensorRevLookup = {
-        '00': 'lux1310',    # Chronos 1.4 original hardware.
-        '14': 'lux1310',    # Chronos 1.4 new production.
-        '21': 'lux2100'     # Chronso 2.1
-    }
-
-    # Do argument parsing
-    parser = argparse.ArgumentParser(description="Chronos control daemon")
-    parser.add_argument('--sensor', metavar='NAME', action='store',
-                        default=sensorRevLookup.get(getBoardRevision()[0:2], 'lux1310'),
-                        help="Override the image sensor driver to use")
-    parser.add_argument('--config', metavar='FILE', action='store',
-                        default='/var/camera/apiConfig.json',
-                        help="Configuration file path")
-    parser.add_argument('--debug', default=False, action='store_true',
-                        help="Enable debug logging")
-    parser.add_argument('--pdb', default=False, action='store_true',
-                        help="Drop into a python debug console on exception")
-    args = parser.parse_args()
-
-    if not args.debug:
-        logging.getLogger().setLevel(logging.INFO)
-    else:
-        logging.getLogger().setLevel(logging.DEBUG)
-
-    # Install exception handlers for interactive debug on exception.
-    if args.pdb:
-        def excepthook(t,v,tb):
-            pdb.traceback.print_exception(t, v, tb)
-            pdb.post_mortem(t=tb)
-        sys.excepthook = excepthook
-        dbg, brk = pdb.set_trace, pdb.set_trace #convenience debugging
-
-        #Fix system not echoing keystrokes in debugger, after restart.
-        try:
-            os.system('stty sane')
-        except Exception:
-            pass
-
-    # Use the GLib mainloop.    
-    mainloop = GLib.MainLoop()
-
-    # Import the desired sensor class.
-    sensorModule = __import__("pychronos.sensors", globals(), None, fromlist=[args.sensor])
-    sensorClass = sensorModule.__getattribute__(args.sensor)
-
-    # Create the camera and objects.
-    sensor = sensorClass()
-    cam  = camera(sensor)
-
-    # Install a timer for battery data monitoring
-    timer = cameraTimer(cam)
-    GLib.timeout_add(1000, timer)
-   
-    name = dbus.service.BusName('ca.krontech.chronos.control', bus=bus)
-    obj  = controlApi(bus, '/ca/krontech/chronos/control', mainloop, cam, configFile=args.config)
-
-    # Run the mainloop.
-    logging.info("Running control service...")
-    mainloop.run()
-
-    # Handle reboot
-    logging.info("Stopping control service... %s", obj.rebootMode)
-    if obj.rebootMode.get('settings', False):
-        os.remove(args.config)
-    
-    if obj.rebootMode.get('power', False):
-        os.system("shutdown -hr now")
-    elif obj.rebootMode.get('reload', True):
-        os.kill(os.getpid(), signal.SIGHUP)
