+++ debian/chronos-control.service
@@ -5,7 +5,7 @@ Requires=chronos-fpga.service chronos-pwrutil.service
 
 [Service]
 Type=dbus
-ExecStart=/usr/bin/cam_control
+ExecStart=/usr/bin/camControl
 BusName=ca.krontech.chronos.control
 RestartForceExitStatus=SIGHUP
 StandardOutput=journal+console
